requirements.txt
discord.py>=2.3.2
tabulate>=0.9.0
python-dotenv>=1.0.1